fix editable
